:truck: api: move ComplementaryCertificationCourseResultForJuryCertification to src
