+++ api/lib/infrastructure/repositories/jury-certification-repository.js
@@ -2,10 +2,10 @@ import _ from 'lodash';
 
 import { knex } from '../../../db/knex-database-connection.js';
 import { JuryCertification } from '../../../src/certification/session-management/domain/models/JuryCertification.js';
+import { ComplementaryCertificationCourseResultForJuryCertification } from '../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
 import { ComplementaryCertificationCourseResultForJuryCertificationWithExternal } from '../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertificationWithExternal.js';
 import { CertificationIssueReport } from '../../../src/certification/shared/domain/models/CertificationIssueReport.js';
 import { NotFoundError } from '../../../src/shared/domain/errors.js';
-import { ComplementaryCertificationCourseResultForJuryCertification } from '../../../src/shared/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
 
 const get = async function (certificationCourseId) {
   const juryCertificationDTO = await _selectJuryCertifications()
