+++ api/src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js
