+++ api/tests/tooling/domain-builder/factory/build-complementary-certification-course-result-for-certification.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertificationCourseResultForJuryCertification } from '../../../../src/shared/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
+import { ComplementaryCertificationCourseResultForJuryCertification } from '../../../../src/certification/session-management/domain/read-models/ComplementaryCertificationCourseResultForJuryCertification.js';
 
 const buildComplementaryCertificationCourseResultForJuryCertification = function ({
   id = 1234,
